requirements.txt
pandas
numpy
matplotlib
seaborn
folium
googlemaps
requests
python-dotenv
geopy